// apps/helpdesk/src/pages/integrations/integration/types.ts
export enum Tab {
    EmailOnboarding = 'onboarding',
    EmailForwarding = 'forwarding',
    EmailVerification = 'verification',
    EmailDomainVerification = 'dns',
    EmailOutboundVerification = 'outbound-verification',
    FacebookCustomerChat = 'customer_chat',
    Preferences = 'preferences',
    FacebookAds = 'ads',
    HttpEvents = 'events',
    Installation = 'installation',
    QuickReplies = 'quick_replies',
    Campaigns = 'campaigns',
    Appearance = 'appearance',
    PhoneVoicemail = 'voicemail',
    PhoneGreetingsMusic = 'greetings-music',
    PhoneIvr = 'ivr',
    CreateWizard = 'create-wizard',
    Languages = 'languages',
    Automate = 'automate',
}